wip: fix clippy warnings
